\name{allelic.kinetics.obj}
\alias{allelic.kinetics.obj}
\docType{data}
\title{
Pre-stored allele-specific bursting kinetic parameters inferred by SCALE.
}
\description{
Pre-computed allele-specific kinetics for demo purpose.
}
\usage{data(allelic.kinetics.obj)}
\value{
List of parameters for allele-specific transcriptional bursting.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(allelic.kinetics.obj)
}
\keyword{datasets}