\name{diff.allelic.obj}
\alias{diff.allelic.obj}
\docType{data}
\title{
Hypothesis testing on two alleles sharing different bursting kinetics.
}
\description{
Pre-stored hypothesis testing results for demo purpose.
}
\usage{data(diff.allelic.obj)}
\value{
List of p-values from testing different allelic burst frequency and burst size.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(diff.allelic.obj)
}
\keyword{datasets}