\name{gene.class.obj}
\alias{gene.class.obj}
\docType{data}
\title{
Bayes framework for gene classification by allele-specific expression data at
single-cell resolution.
}
\description{
Pre-stored classification results for demo purpose.
}
\usage{data(gene.class.obj)}
\value{
List of gene categorization results for the demo dataset.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(gene.class.obj)
}
\keyword{datasets}