\name{mouse.blastocyst}
\alias{mouse.blastocyst}
\docType{data}
\title{
Demo dataset of single-cell RNA sequencing of 122 mouse blastocyst cells.
}
\description{
Pre-stored single-cell RNA sequencing dataset of 122 mouse blastocyst cells.
}
\usage{data(mouse.blastocyst)}
\value{
List of demo dataset of single-cell RNA sequencing of mouse blastocyst cells.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(mouse.blastocyst)
}
\keyword{datasets}