\name{scale-package}
\alias{scale-package}
\alias{scale}
\docType{package}
\title{
\packageTitle{scale}
}
\description{
\packageDescription{scale}
}
\details{
scale provides a reference implementation of the Scaleable Langevin
Exact Algorithm as described in Pollock et al. (2016).

Note that this implementation attempts to explicitly replicate the
mathematical formulation of the algorithm and computational efficiency
has been sacrificed to a considerable degree in order to maintain the
relationship between the mathematical presentation and the
implementation itself.

# Examples
This section explains how the examples presented in Pollock et
al. (2016) can be reproduced.

## Preliminaries
In order to replicate the examples presented in the manucript the
following preliminary steps are required:

 1. Download the relevant data files from
 (http://www.warwick.ac.uk/mpollock/scale)
 
 2. Download the precomputed centering and preconditioning data from
 (http://www.warwick.ac.uk/mpollock/scale)

 ## Logistic Regression With Limited Informatioin

 ## Logistic Regression for Airline Data

 ## Large Logistic Regression

 ## Contaminated Regression
 
}
\author{
\packageAuthor{scale}

Maintainer: \packageMaintainer{scale}
}
\references{
  Pollock, Murray, Paul Fearnhead, Adam M. Johansen, and Gareth O. Roberts. "The scalable Langevin exact algorithm: Bayesian inference for big data." arXiv preprint arXiv:1609.03436 (2016).

}
\keyword{ package }
\seealso{
}
\examples{
}
