% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scaler.R
\name{Scaler}
\alias{Scaler}
\title{Scaler'}
\arguments{
\item{parameters}{A data.frame containing the following columns:
\itemize{
\item \code{x} - The maximum x value
\item \code{y} - The maximum y value
}}
}
\value{
a list containing two scaling utilities:
\itemize{
\item \code{scale <- parameters |> Scaler()}
\item \code{x |> scale[['X']]()}
\item \code{y |> scale[['Y']]()}
}
}
\description{
Scale X and Y Coordinates given a maximum x and y value passed as parameters
}
